requirements.txt

django
pillow
django-admin-interface
django-crispy-forms
mysqlclient
pymysql
djangorestframework
requests
mysqlclient-1.4.6-cp38-cp38-win32.whl